renamed Enum.ts
